Create questions.md
Created a file which will be cloned with the repo for students to write the answers to the lab questions in. This will hopefully speed up the lab by giving students a place to write answers without needing to create a Google doc, new file, etc. Hopefully it will also make more clear the expectation that students respond to the questions and submit their response.